rename package changer to changers
